// frontend/src/types/enums.ts
export enum Perfil {
  ADMINISTRADOR = 'ADMINISTRADOR',
  TECNICO_PATRIMONIO = 'TECNICO_PATRIMONIO',
  CHEFE_SETOR = 'CHEFE_SETOR',
  USUARIO_CONSULTA = 'USUARIO_CONSULTA',
}

export enum EstadoConservacao {
  EXCELENTE = 'EXCELENTE',
  BOM = 'BOM',
  REGULAR = 'REGULAR',
  RUIM = 'RUIM',
  PESSIMO = 'PESSIMO',
}

export enum StatusItem {
  ATIVO = 'ATIVO',
  INATIVO = 'INATIVO',
  EM_MOVIMENTACAO = 'EM_MOVIMENTACAO',
  BAIXADO = 'BAIXADO',
  EM_MANUTENCAO = 'EM_MANUTENCAO',
}

export enum StatusMovimentacao {
  AGUARDANDO_CONFIRMACAO_ENTREGA = 'AGUARDANDO_CONFIRMACAO_ENTREGA',
  AGUARDANDO_CONFIRMACAO_RECEBIMENTO = 'AGUARDANDO_CONFIRMACAO_RECEBIMENTO',
  AGUARDANDO_APROVACAO_PATRIMONIO = 'AGUARDANDO_APROVACAO_PATRIMONIO',
  APROVADA = 'APROVADA',
  REJEITADA = 'REJEITADA',
  CONCLUIDA = 'CONCLUIDA',
  CANCELADA = 'CANCELADA',
}

export enum TipoEntrada {
  COMPRA = 'COMPRA',
  DOACAO = 'DOACAO',
  CESSAO = 'CESSAO',
  TRANSFERENCIA = 'TRANSFERENCIA',
  PRODUCAO_PROPRIA = 'PRODUCAO_PROPRIA',
  OUTRO = 'OUTRO',
}

export enum MotivoBaixa {
  LEILAO = 'LEILAO',
  DOACAO = 'DOACAO',
  INSERVIVEL = 'INSERVIVEL',
  EXTRAVIO = 'EXTRAVIO',
  DESCARTE = 'DESCARTE',
  TRANSFERENCIA_DEFINITIVA = 'TRANSFERENCIA_DEFINITIVA',
  OUTRO = 'OUTRO',
}

export enum StatusInventario {
  ABERTO = 'ABERTO',
  CONCLUIDO = 'CONCLUIDO',
}

export enum StatusInventarioItem {
  PENDENTE = 'PENDENTE',
  LOCALIZADO = 'LOCALIZADO',
  NAO_LOCALIZADO = 'NAO_LOCALIZADO',
}